# openapi.yaml
openapi: 3.1.0
info:
  title: Affilite Mix API
  version: "1"
  description: "Multi-tenant affiliate marketing platform API. Routes marked x-internal: true are not part of the public consumer contract."
  contact:
    name: Affilite Mix
  license:
    name: Proprietary
servers:
  - url: "https://compareai.site"
    description: Production (compareai.site)
  - url: "https://staging.compareai.site"
    description: Staging
paths:
...
  /api/admin/upload/finalize:
    post:
      summary: POST /api/admin/upload/finalize (site) — auth=admin
      operationId: postApiAdminUploadFinalize
      tags:
        - Admin
      x-auth: admin
      x-adminRequired: true
      x-tenantScope: site
      x-rateLimit: true
      x-csrf: true
      x-requestSchema: UploadFinalizeInput
      x-responseSchema: UploadedImage
      x-notes: null
      security:
        - bearerAuth: []
      x-internal: true
      responses:
        200:
          description: Successful response (UploadedImage)
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        401:
          description: Unauthorized
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        403:
          description: Forbidden — insufficient role or cross-tenant access
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
...
  /api/auth/forgot-password:
    post:
      summary: POST /api/auth/forgot-password (global) — public
      operationId: postApiAuthForgotpassword
      tags:
        - Auth
      x-auth: public
      x-adminRequired: false
      x-tenantScope: global
      x-rateLimit: true
      x-csrf: true
      x-requestSchema: "{ email: string; turnstileToken: string }"
      x-responseSchema: Ok (always 200 to avoid enumeration)
      x-notes: "Audit-logged as `request_password_reset`; throttled per IP + email."
      security: []
      responses:
        200:
          description: Successful response
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
          content:
            application/json:
              schema:
                type: object
        429:
          description: Rate limited
          headers:
            API-Version:
              $ref: "#/components/headers/API-Version"
...
components:
...
  headers:
    API-Version:
      description: Numeric major version of the served API contract.
      schema:
        type: string
        pattern: "^[1-9][0-9]*$"
        example: "1"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...